refactor(FR-3683): fold `dom.ts` back into `FormStore.ts`
`isVisible` and the reduced-motion-aware `scrollIntoView` were split out
when both `<Form>` and the store needed them; the store is the only
caller of the latter and the natural owner of both, so they live there
and `Form.tsx` imports `isVisible` from it. No behaviour change.
